added Commnets button and coding
