+++ app/src/main/java/com/team03/coursewarehub/Youtube/VideoDisplay.java
@@ -39,9 +39,6 @@ public class VideoDisplay extends YouTubeBaseActivity implements YouTubePlayer.O
     private static final int RECOVERY_DIALOG_REQUEST = 10;
     public static final String API_KEY = "AIzaSyC29Cn2stiksjnP33HeegXMNCCAoMRzgnc";
     String videoId;
-    List<String> CommentList = new ArrayList<String>();
-    ArrayAdapter<String> adapter;
-    String email_pattern = "[a-zA-Z0-9._-]+@[a-z]+\\.+[a-z]+";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -72,53 +69,6 @@ public class VideoDisplay extends YouTubeBaseActivity implements YouTubePlayer.O
         final TextView tv = (TextView)findViewById(R.id.textView);
         tv.setText(name);
 
-        //Get the selected position from the previous activity
-        final int iPosition = intent.getIntExtra("Position", -1);
-
-        BootstrapButton bbp = (BootstrapButton) findViewById(R.id.btnPost);
-        BootstrapButton bbc = (BootstrapButton) findViewById(R.id.btnCancel);
-        final BootstrapEditText betComment = (BootstrapEditText) findViewById(R.id.commentEditText);
-        final BootstrapEditText betUser = (BootstrapEditText) findViewById(R.id.usernameEditText);
-        final BootstrapEditText betEmail = (BootstrapEditText) findViewById(R.id.emailEditText);
-
-
-        bbp.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-
-                //input validation
-                if(!CheckValidation(betUser.getText().toString(), betComment.getText().toString(), betEmail.getText().toString()))
-                {
-                    return; // validation was wrong, exit this onClick method
-                }
-
-                //prepare data to set in Firebase
-                Comments comment = new Comments (betUser.getText().toString(), betComment.getText().toString(), betEmail.getText().toString());
-
-                //get Firebase reference, then set value to Firebase
-                Firebase refFB = ref.child("Videos").child(String.valueOf(iPosition)).child("Comment");
-                refFB.push().setValue(comment);
-
-                // Clear the text fields
-                betComment.setText("");
-                betUser.setText("");
-                betEmail.setText("");
-
-                //refresh this page
-                Intent intent = getIntent();
-                finish();
-                startActivity(intent);
-            }
-        });
-        bbc.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                betComment.setText("");
-                betUser.setText("");
-                betEmail.setText("");
-            }
-        });
-
         ref.addChildEventListener(new ChildEventListener() {
             @Override
             public void onChildAdded(DataSnapshot dataSnapshot, String s) {
@@ -127,33 +77,6 @@ public class VideoDisplay extends YouTubeBaseActivity implements YouTubePlayer.O
                     // Assigning Image to Image view
                     UrlImageViewHelper.setUrlDrawable(imgHeader, tempHeaderImage);
                 }
-                if (dataSnapshot.getKey().toString() == "Videos") {
-                    // display list to listview.
-                    for (DataSnapshot snapshot : dataSnapshot.getChildren()) {
-                        if (iPosition == Integer.parseInt(snapshot.getKey().toString())) {
-                            //Comment shows only selected the video from the previous activity
-                            if (snapshot.child("Comment").exists()) {
-                                // to avoid reset when "Comment" does not exist in layers
-                                List<Comments> com = new ArrayList<Comments>();
-                                for (DataSnapshot ds : snapshot.child("Comment").getChildren()) {
-                                // this loop gets from the oldest comment to the latest comment
-                                        com.add(ds.getValue(Comments.class));
-                                }
-                                for(long i = (snapshot.child("Comment").getChildrenCount() - 1) ; i >= 0 ; i--){
-                                //  keep comments in ArrayList, then display reverse order
-                                    String strText = com.get((int)i).getText();
-                                    String strUser = com.get((int)i).getName();
-                                    String strComment = strText + " by " + strUser;
-                                    CommentList.add(strComment);
-                                }
-                                ListView lv = (ListView) findViewById(R.id.listViewComment);
-                                lv.setAdapter(adapter);
-                                setListViewHeightBasedOnItems(lv);
-                            }
-                        }
-                    }
-                }
-
             }
 
             @Override
@@ -181,76 +104,8 @@ public class VideoDisplay extends YouTubeBaseActivity implements YouTubePlayer.O
         YouTubePlayerView youTubeView = (YouTubePlayerView) findViewById(R.id.youtube_view);
         youTubeView.initialize(API_KEY, this);
 
-        ListView lv = (ListView) findViewById(R.id.listViewComment);
-        adapter = new ArrayAdapter<String>(this, R.layout.main_search_list_item, R.id.product_name, CommentList);
-        lv.setAdapter(adapter);
-        setListViewHeightBasedOnItems(lv);
     }
 
-    public boolean CheckValidation(String strName, String strText, String strEmail)
-    {
-        if (strText.trim().equals(""))
-        {
-            Toast.makeText(this, "Input your comment", Toast.LENGTH_LONG).show();
-            return false;
-        }
-        if (strName.trim().equals(""))
-        {
-            Toast.makeText(this, "Input your name", Toast.LENGTH_LONG).show();
-            return false;
-        }
-        if (strEmail.trim().equals(""))
-        {
-            Toast.makeText(this, "Input your email address", Toast.LENGTH_LONG).show();
-            return false;
-        }
-        if(!strEmail.matches(email_pattern))
-        {
-            Toast.makeText(this, "Email format is wrong", Toast.LENGTH_LONG).show();
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * Sets ListView height dynamically based on the height of the items.
-     *
-     * @param listView to be resized
-     * @return true if the listView is successfully resized, false otherwise
-     */
-    public static boolean setListViewHeightBasedOnItems(ListView listView) {
-
-        ListAdapter listAdapter = listView.getAdapter();
-        if (listAdapter != null) {
-
-            int numberOfItems = listAdapter.getCount();
-
-            // Get total height of all items.
-            int totalItemsHeight = 0;
-            for (int itemPos = 0; itemPos < numberOfItems; itemPos++) {
-                View item = listAdapter.getView(itemPos, null, listView);
-                item.measure(0, 0);
-                //item.measure(View.MeasureSpec.makeMeasureSpec(0, View.MeasureSpec.UNSPECIFIED), View.MeasureSpec.makeMeasureSpec(0, View.MeasureSpec.UNSPECIFIED));
-                totalItemsHeight += item.getMeasuredHeight();
-            }
-
-            // Get total height of all item dividers.
-            int totalDividersHeight = listView.getDividerHeight() *
-                    (numberOfItems - 1);
-
-            // Set list height.
-            ViewGroup.LayoutParams params = listView.getLayoutParams();
-            params.height = totalItemsHeight + totalDividersHeight;
-            listView.setLayoutParams(params);
-            listView.requestLayout();
-            return true;
-
-        } else {
-            return false;
-        }
-
-    }
 //    public String extractYTId(String ytUrl) {
 //        //String pattern = "(?<=watch\\?v=|/videos/|embed\\/)[^#\\&\\?]*";
 //        String pattern = "(?<=watch\\?v=|/videos/|embed\\/|youtu.be\\/|\\/v\\/|watch\\?v%3D|%2Fvideos%2F|embed%\u200C\u200B2F|youtu.be%2F|%2Fv%2F)[^#\\&\\?\\n]*";
@@ -403,30 +258,5 @@ public class VideoDisplay extends YouTubeBaseActivity implements YouTubePlayer.O
 
         }
 
-    }
-
-    public static class Comments {
-        String name;
-        String text;
-        String email;
-
-        public Comments(){
-        }
-
-        public Comments(String name, String message, String email){
-            this.name = name;
-            this.text = message;
-            this.email = email;
-        }
-
-        public String getName(){
-            return name;
-        }
-        public String getText(){
-            return text;
-        }
-        public String getEmail(){
-            return email;
-        }
     }
 }
