+++ app/src/main/res/layout/activity_video_display.xml
@@ -50,96 +50,6 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"/>
 
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceMedium"
-            android:text="Comments:"
-            android:id="@+id/textView2" />
-
-        <!--<com.beardedhen.androidbootstrap.BootstrapEditText-->
-            <!--android:layout_width="match_parent"-->
-            <!--android:layout_height="wrap_content"-->
-            <!--android:inputType="textMultiLine"-->
-            <!--android:id="@+id/editText"-->
-            <!--app:roundedCorners="true" />-->
-        <ListView
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/listViewComment" />
-        <TableRow
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginRight="18dp"
-                android:text="Comment:"/>
-            <com.beardedhen.androidbootstrap.BootstrapEditText
-                android:layout_width="fill_parent"
-                android:inputType="textMultiLine"
-                android:id="@+id/commentEditText"
-                app:roundedCorners="true"/>
-        </TableRow>
-        <TableRow
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginRight="10dp"
-                android:text="User Name:"/>
-            <com.beardedhen.androidbootstrap.BootstrapEditText
-                android:layout_width="fill_parent"
-                android:inputType="textMultiLine"
-                android:id="@+id/usernameEditText"
-                app:roundedCorners="true"/>
-        </TableRow>
-        <TableRow
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginRight="46dp"
-                android:text="Email:"/>
-            <com.beardedhen.androidbootstrap.BootstrapEditText
-                android:layout_width="fill_parent"
-                android:inputType="textMultiLine"
-                android:id="@+id/emailEditText"
-                app:roundedCorners="true"/>
-        </TableRow>
-        <TableRow
-            android:layout_width="match_parent"
-            android:layout_height="68dp">
-
-            <com.beardedhen.androidbootstrap.BootstrapButton
-                android:id="@+id/btnPost"
-                android:layout_height="wrap_content"
-                android:layout_width="wrap_content"
-                android:layout_marginTop="20dp"
-                android:text="Post"
-                app:bootstrapBrand="success"
-                app:bootstrapSize="lg"
-                app:buttonMode="regular"
-                app:roundedCorners="true"
-                app:showOutline="false"
-                />
-
-            <com.beardedhen.androidbootstrap.BootstrapButton
-                android:id="@+id/btnCancel"
-                android:layout_height="wrap_content"
-                android:layout_width="wrap_content"
-                android:layout_marginTop="20dp"
-                android:text="Cancel"
-                app:bootstrapBrand="success"
-                app:bootstrapSize="lg"
-                app:buttonMode="regular"
-                app:roundedCorners="true"
-                app:showOutline="false" />
-        </TableRow>
-
-
     </LinearLayout>
     </ScrollView>
